pyproject.toml
[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "bme280"
version = "1.0.0"
description = "A python library for accessing the BME280 combined humidity and pressure sensor from Bosch."
authors = ["Christian Nicolai"]
license = "MIT"
homepage = "https://dev.mycrobase.de/gitea/cn/python-bme280"

[tool.poetry.dependencies]
python = "^3.7"
smbus2 = "*"
mpu6050-raspberrypi = "^1.2"
anvil-uplink = "^0.3.41"

[tool.poetry.dev-dependencies]
mock = "*"
pylint = "*"
pytest = "*"
snapshottest = "*"
rope = "*"
yapf = "*"
